Refactoring
- project cleanup and restructuring
- implement MVVM
