--- BarcodeScanner/BarcodeScanner/BarcodeScannerView.swift
@@ -1,60 +1,0 @@
-//
-//  ContentView.swift
-//  BarcodeScanner
-//
-//  Created by Norbert Agoston on 15.02.2024.
-//
-
-import SwiftUI
-
-struct AlertItem: Identifiable {
-    let id = UUID()
-    let title: String
-    let message: String
-    let dismissButton: Alert.Button
-}
-
-struct AlertContext {
-    static let invalidDeviceInput = AlertItem(title: "Invalid Device Input", 
-                                              message: "Something is wrong with the camera.",
-                                              dismissButton: .default(Text("Ok")))
-    
-    static let invalidScannedType = AlertItem(title: "Invalid Scan Type ",
-                                              message: "Scanned value is not valid.",
-                                              dismissButton: .default(Text("Ok")))
-}
-
-struct BarcodeScannerView: View {
-    @State private var scannedCode = ""
-    @State private var alertItem: AlertItem?
-    
-    var body: some View {
-        NavigationView {
-            VStack {
-                ScannerView(scannedCode: $scannedCode,
-                            alertItem: $alertItem)
-                    .frame(maxWidth: .infinity, maxHeight: 300)
-                Spacer().frame(height: 60)
-                Label("Scanned Barcode:", systemImage: "barcode.viewfinder")
-                    .font(.title)
-                Text(scannedCode.isEmpty ? "Not yet scaned" : scannedCode)
-                    .bold()
-                    .font(.largeTitle)
-                    .foregroundColor(scannedCode.isEmpty ? .red : .green)
-                    .padding()
-            }
-            .navigationTitle("Barcode Scanner")
-            .alert(item: $alertItem) { alertItem in
-                Alert(title: Text(alertItem.title),
-                      message: Text(alertItem.message),
-                      dismissButton: alertItem.dismissButton
-                      
-                )
-            }
-        }
-    }
-}
-
-#Preview {
-    BarcodeScannerView()
-}
