+++ BarcodeScanner/BarcodeScanner/Screens/BarcodeScanner/BarcodeScannerViewModel.swift
@@ -1,0 +1,21 @@
+//
+//  BarcodeScannerViewModel.swift
+//  BarcodeScanner
+//
+//  Created by Norbert Agoston on 15.02.2024.
+//
+
+import SwiftUI
+
+final class BarcodeScannerViewModel: ObservableObject {
+    @Published var scannedCode = ""
+    @Published var alertItem: AlertItem?
+    
+    var statusText: String {
+        scannedCode.isEmpty ? "Not yet scaned" : scannedCode
+    }
+    
+    var statusTextColor: Color {
+        scannedCode.isEmpty ? .red : .green
+    }
+}
