+++ BarcodeScanner/BarcodeScanner/Views/UIKitComponents/ScannerVC.swift
