+++ BarcodeScanner/BarcodeScanner/Views/UIKitComponents/ScannerView.swift
